Test when an inherited virtual member is overridden and listed in a derived class's interface definition
This was causing problems because the method is not repeated in the primary vtable (it shows up in the position expected by the base class), but CppAbi does not know that it is inherited until after the impl has been emitted, so it emits the index that it is listed in the derived class.

Note that the generator only does this with virtual destructors (since we never call these virtually).