Removed property accessors ignored checking code.
The logic here is just wrong... first even if one of the accessors is ignored that does not mean we should ignore the whole property. Additionally, the code logic to check if an accessor is ignored is faulty, we can not rely on the return value of the Visit functions, but check if the declaration itself was ignored.