build: use LLVM from an install directory
Require the LLVM to be installed into deps/llvm/install, and use the
LLVM from there. This simplifies the include and library paths. It also
resembles more the way things should work if the LLVM was provided by
the distribution.

We still need some hacks to include the private clang CodeGen headers,
though.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@iki.fi>