Remove PrimitiveTypeExtensions.ConvertToType
C++ types cannot be converted to C# types without knowing details of the
target (e.g. what is the size of "long") and
PrimitiveTypeExtensions.ConvertToType does not handle that correctly.

However, ConvertToType is not used anywhere, so let's just remove it so
that no one will use the method.

We could also fix the method, but it's an extension method for
PrimitiveType, and I don't think such an extension method should be used
for the task, as you would also need to pass it information about the
target system.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@iki.fi>