Remove PrimitiveTypes Int8 and UInt8
As enum PrimitiveType is supposed to represent C++ types, having Int8
and UInt8 there is not correct.

This patch removes those values. As Int8 and UInt8 were set to be
equivalent to Char and UChar, we can just change all references to Int8
and UInt8 to Char and UChar.

This doesn't change the behavior, but makes it clearer that the types
for 'char' and 'unsigned char' are not necessarily 8 bits.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@iki.fi>