Pass Driver to CSharpTypePrinter
We need Driver.TargeInfo in CSharpTypePrinter so that we can find out the
widths of the integer types.

This patch changes CSharpTypePrinter to get the Driver reference in its
constructor. As all the other constructor parameters can be found from the
Driver reference, we can remove all the other parameters.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@iki.fi>