Parser/CppParser: Use installed LLVM config.h
The parsers include clang's private config.h file, which is not present
in the installed LLVM. However, the installed LLVM has
llvm/Config/config.h, which contains the same things that clang's
private config.h contains.

So use the LLVM's public config.h instead, so that we're more ready to
use a real LLVM installation.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@iki.fi>