build: remove explicit platforms
Selecting the platform when building feels a bit useless.

On Windows, the LLVM build is (afaik) always 32-bit. That means that all
the C++ libs need to be 32-bit, and in turn all the C# libs/exes using
those native libraries need to be 32-bit.

On Linux, the LLVM is in the system's native bitness. So if your system
is 64-bit, there's no reason for you to try to compile 32-bit CppSharp.

All in all, things seem cleaner with this patch: remove the platforms,
use 32-bit architecture on windows, and the native architecture on
others.

This patch also changes the configurations order so that Release is
first. That makes Release the default target, making it easier for the
user to compile CppSharp.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@iki.fi>