Add -fpermissive to gcc cflags
We again have code that does not compile with gcc:

../../../src/CppParser/AST.h:770:19: error: declaration of
‘CppSharp::CppParser::AST::MacroLocation
CppSharp::CppParser::AST::PreprocessedEntity::MacroLocation’
[-fpermissive]
     MacroLocation MacroLocation;
                   ^
../../../src/CppParser/AST.h:756:12: error: changes meaning of
‘MacroLocation’ from ‘enum class
CppSharp::CppParser::AST::MacroLocation’ [-fpermissive]
 enum class MacroLocation

Instead of trying to figure out a good name for the property, and to
avoid the same issues in the future, let's just add the -fpermissive
flag back.

Maybe it can be removed in the future if and when Linux support works
well and is supported by the build test bot.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@iki.fi>