Stored specializations in the templated class itself because ClassTemplate is not a complete declaration.
ClassTemplate in Clang does not actually represent a complete declaration. For example, let's see the implementation of clang::ClassTemplateDecl::isThisDeclarationADefinition:

    bool isThisDeclarationADefinition() const {
      return getTemplatedDecl()->isThisDeclarationADefinition();
    }

(http://clang.llvm.org/doxygen/DeclTemplate_8h_source.html#l01995 at the time of this commit)

The above shows that class templates do not have definitions, instead, their "definitions" are those of the real templates classes.
In addition let's see ClassTemplateDecl::getCanonicalDecl():

    const ClassTemplateDecl *getCanonicalDecl() const {
      return cast<ClassTemplateDecl>(
               RedeclarableTemplateDecl::getCanonicalDecl());
    }

(http://clang.llvm.org/doxygen/DeclTemplate_8h_source.html#l02023 at the time of this commit)

In turn RedeclarableTemplateDecl::getCanonicalDecl() is defined as:

   RedeclarableTemplateDecl *getCanonicalDecl() override {
     return getFirstDecl();
   }

where getFirstDecl() returns just that - the first encountered declaration which might as well be a forward one.
This means that the only complete declaration ClassTemplateDecl can point to is the template class itself,
and the latter is therefore our only choice if we want to have a full list of all specializations of that template class.

Signed-off-by: Dimitar Dobrev <dpldobrev@protonmail.com>