fixes #1191 CppSharp.CLI.exe --rtti sets -fno-rtti to clang
The problem being fixed is that although --rtti is given as argument to
CppSharp.CLI.exe the runtime type information is turned off. The error
is happening because the Generator options are copied over to the
parserOptions after the parserOptions have been used to generate the
cmake call. Thus, the default value of RTTI (off) is always evaluated
and cmake will always get the parameter -fno-rtti.
The same issue seems to apply to the unity build.

This patch fixes the issue by moving the 2 assignments of Generator
options properties into ParserOptions properties before they are
evaluated in parserOptions.Setup();