Always use the mangled names as reported by Clang
The names in the binaries themselves are sometimes different than those Clang reports. However, it seems that the reported ones are the actually correct for invocation. For example, the name passed to dlsym() (for variables) must not be prepended with an underscore (https://developer.apple.com/library/archive/documentation/System/Conceptual/ManPages_iPhoneOS/man3/dlsym.3.html). Another example, again on macOS, is with extern C functions which sometimes also receive an _ at the front.

Signed-off-by: Dimitar Dobrev <dpldobrev@protonmail.com>